Add logic for queing messages to neighbors
Earlier I was trying to send heartbeat and log messages simultaneously. It was causing a lot ClosedChannelExcpetions. I resolved it with this queue. However I still have a problem with serialization of LogEntry messages. GetValue works but SetValue and RemoveValue (needing LogEntry and CommitEntry messages sent between servers) does not.